Move QuickSort classes to a fitting package
They were in ...memory, now in ...internal.util.sort
